Separate VgMarkConfig from MarkConfig
