+++ src/mark.ts
@@ -1,4 +1,6 @@
 import {toSet} from './util';
+import {Orient, Interpolate, VgMarkConfig} from './vega.schema';
+export {Orient} from './vega.schema';
 
 export namespace Mark {
   export const AREA: 'area' = 'area';
@@ -91,23 +93,6 @@ export function isPrimitiveMark(mark: string | MarkDef): mark is Mark {
   return markType in PRIMITIVE_MARK_INDEX;
 }
 
-export type FontStyle = 'normal' | 'italic';
-export type FontWeight = 'normal' | 'bold';
-/**
- * @TJS-type integer
- * @minimum 100
- * @maximum 900
- */
-export type FontWeightNumber = number;
-export type HorizontalAlign = 'left' | 'right' | 'center';
-export type Interpolate = 'linear' | 'linear-closed' |
-  'step' | 'step-before' | 'step-after' |
-  'basis' | 'basis-open' | 'basis-closed' |
-  'cardinal' | 'cardinal-open' | 'cardinal-closed' |
-  'bundle' | 'monotone';
-export type Orient = 'horizontal' | 'vertical';
-export type VerticalAlign = 'top' | 'middle' | 'bottom';
-
 export const STROKE_CONFIG = ['stroke', 'strokeWidth',
   'strokeDash', 'strokeDashOffset', 'strokeOpacity'];
 
@@ -115,7 +100,7 @@ export const FILL_CONFIG = ['fill', 'fillOpacity'];
 
 export const FILL_STROKE_CONFIG = [].concat(STROKE_CONFIG, FILL_CONFIG);
 
-export interface MarkConfig {
+export interface MarkConfig extends VgMarkConfig {
 
   // ---------- Color ----------
   /**
@@ -130,169 +115,6 @@ export interface MarkConfig {
    * Default color.
    */
   color?: string;
-
-  /**
-   * Default Fill Color.  This has higher precedence than config.color
-   */
-  fill?: string;
-
-  /**
-   * Default Stroke Color.  This has higher precedence than config.color
-   */
-  stroke?: string;
-
-  // ---------- Opacity ----------
-  /**
-   * @minimum 0
-   * @maximum 1
-   */
-  opacity?: number;
-
-
-  /**
-   * @minimum 0
-   * @maximum 1
-   */
-  fillOpacity?: number;
-
-  /**
-   * @minimum 0
-   * @maximum 1
-   */
-  strokeOpacity?: number;
-
-  // ---------- Stroke Style ----------
-  /**
-   * @minimum 0
-   */
-  strokeWidth?: number;
-
-  /**
-   * An array of alternating stroke, space lengths for creating dashed or dotted lines.
-   */
-  strokeDash?: number[];
-
-  /**
-   * The offset (in pixels) into which to begin drawing with the stroke dash array.
-   */
-  strokeDashOffset?: number;
-
-  // ---------- Orientation: Bar, Tick, Line, Area ----------
-  /**
-   * The orientation of a non-stacked bar, tick, area, and line charts.
-   * The value is either horizontal (default) or vertical.
-   * - For bar, rule and tick, this determines whether the size of the bar and tick
-   * should be applied to x or y dimension.
-   * - For area, this property determines the orient property of the Vega output.
-   * - For line, this property determines the sort order of the points in the line
-   * if `config.sortLineBy` is not specified.
-   * For stacked charts, this is always determined by the orientation of the stack;
-   * therefore explicitly specified value will be ignored.
-   */
-  orient?: Orient;
-
-  // ---------- Interpolation: Line / area ----------
-  /**
-   * The line interpolation method to use for line and area marks. One of the following:
-   * - `"linear"`: piecewise linear segments, as in a polyline.
-   * - `"linear-closed"`: close the linear segments to form a polygon.
-   * - `"step"`: alternate between horizontal and vertical segments, as in a step function.
-   * - `"step-before"`: alternate between vertical and horizontal segments, as in a step function.
-   * - `"step-after"`: alternate between horizontal and vertical segments, as in a step function.
-   * - `"basis"`: a B-spline, with control point duplication on the ends.
-   * - `"basis-open"`: an open B-spline; may not intersect the start or end.
-   * - `"basis-closed"`: a closed B-spline, as in a loop.
-   * - `"cardinal"`: a Cardinal spline, with control point duplication on the ends.
-   * - `"cardinal-open"`: an open Cardinal spline; may not intersect the start or end, but will intersect other control points.
-   * - `"cardinal-closed"`: a closed Cardinal spline, as in a loop.
-   * - `"bundle"`: equivalent to basis, except the tension parameter is used to straighten the spline.
-   * - `"monotone"`: cubic interpolation that preserves monotonicity in y.
-   */
-  interpolate?: Interpolate;
-  /**
-   * Depending on the interpolation type, sets the tension parameter (for line and area marks).
-   * @minimum 0
-   * @maximum 1
-   */
-  tension?: number;
-
-  /**
-   * The default symbol shape to use. One of: `"circle"` (default), `"square"`, `"cross"`, `"diamond"`, `"triangle-up"`, or `"triangle-down"`, or a custom SVG path.
-   */
-  shape?: string;
-
-  /**
-   * The pixel area each the point/circle/square.
-   * For example: in the case of circles, the radius is determined in part by the square root of the size value.
-   * @minimum 0
-   */
-  size?: number;
-
-  // Text / Label Mark Config
-
-  /**
-   * The horizontal alignment of the text. One of left, right, center.
-   */
-  align?: HorizontalAlign;
-
-  /**
-   * The rotation angle of the text, in degrees.
-   * @minimum 0
-   * @maximum 360
-   */
-  angle?: number;
-
-  /**
-   * The vertical alignment of the text. One of top, middle, bottom.
-   */
-  baseline?: VerticalAlign;
-
-  /**
-   * The horizontal offset, in pixels, between the text label and its anchor point. The offset is applied after rotation by the angle property.
-   */
-  dx?: number;
-
-  /**
-   * The vertical offset, in pixels, between the text label and its anchor point. The offset is applied after rotation by the angle property.
-   */
-  dy?: number;
-
-  /**
-   * Polar coordinate radial offset, in pixels, of the text label from the origin determined by the x and y properties.
-   * @minimum 0
-   */
-  radius?: number;
-
-  /**
-   * Polar coordinate angle, in radians, of the text label from the origin determined by the x and y properties. Values for theta follow the same convention of arc mark startAngle and endAngle properties: angles are measured in radians, with 0 indicating "north".
-   */
-  theta?: number;
-
-  /**
-   * The typeface to set the text in (e.g., Helvetica Neue).
-   * @minimum 0
-   */
-  font?: string;
-
-  /**
-   * The font size, in pixels.
-   * @minimum 0
-   */
-  fontSize?: number;
-
-  /**
-   * The font style (e.g., italic).
-   */
-  fontStyle?: FontStyle;
-  /**
-   * The font weight (e.g., `"normal"`, `"bold"`, `900`).
-   */
-  fontWeight?: FontWeight | FontWeightNumber;
-
-  /**
-   * Placeholder Text
-   */
-  text?: string;
 }
 
 export const defaultMarkConfig: MarkConfig = {
